Customize bootstrap timing for sidebar toggle
A plotly graph gets resized after the sidebar is toggled. With the ease
timing function, the transition appears to have settled down during the
final part of the transition, and so the plot resize is startling.  With
the ease-in timing function, the transition is still barrelling along at
the moment that the resize happens, and so the resize is less
eye-catching.