Control timing of initial DOM modifications
Plots and tables, along with associated controls and event listeners,
are added to the DOM by a series of .then handlers.  For example,
when the app is loaded, requests to the back end for plots on panes that
are not initially visible are delayed until the plots on the visible
summary pane have been rendered.